+++ website/models.py
@@ -1,14 +1,17 @@
 from django.db import models
 from django.contrib.auth.models import User
 from django.core.validators import MaxValueValidator, MinValueValidator
-
-
-from django.conf import settings
-import stripe
-stripe.api_key = settings.STRIPE_SECRET_KEY
-import datetime
 from django.utils import timezone
+from django.conf import settings
 
+import stripe
+import datetime
+from django_countries.fields import CountryField
+
+#from surveys.models import Product
+
+
+stripe.api_key = settings.STRIPE_SECRET_KEY
 # Create your models here.
 '''
 class SalesArgument(models.Model):
@@ -139,3 +142,23 @@ class Subscriber(models.Model):
     def __str__(self):
         """String for representing the Plan object (in Admin site etc.)."""
         return self.user.username
+
+class Organization(models.Model):
+    owner = models.OneToOneField(User, blank=True, null=True, on_delete=models.CASCADE, help_text='User who owns this Organization')
+    name = models.CharField(max_length=255, blank=True, null=True, help_text='Name of the Organization')
+    #active_products = models.ManyToManyField(Product, blank=True, help_text='Products this organization is currently using')
+    address_line_1 = models.CharField(max_length=255, blank=True, null=True, help_text='Adress of the Organization')
+    address_line_2 = models.CharField(max_length=255, blank=True, null=True, help_text='Address contd.')
+    zip_code =  models.CharField(max_length=255, blank=True, null=True, help_text='Zip code of the Organization')
+    city =  models.CharField(max_length=255, blank=True, null=True, help_text='City where the Organization is located')
+    country =  CountryField(blank=True, null=True, help_text='Country where the Organization is located')
+
+
+    ##Todo
+    #Add support for other Users than owner to be allowed to change organization
+    def __str__(self):
+        """String for representing the Organization object (in Admin site etc.)."""
+        if self.name != None and self.name != '':
+            return self.name
+        else:
+            return "Organization object owned by %s."%(self.owner)
